% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.prob.R
\name{cor.prob}
\alias{cor.prob}
\title{Correlation matrix}
\source{
https://stat.ethz.ch/pipermail/r-help/2001-November/016201.html
}
\usage{
cor.prob(X, dfr = nrow(X) - 2)
}
\description{
Correlation Matrix with strength and significance
}
