% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorCustomMatrix.R
\name{CorMatrixCustom}
\alias{CorMatrixCustom}
\title{Custom Correlation Matrix}
\source{
http://druedin.com/2012/08/11/moving-averages-in-r/
}
\usage{
CorMatrixCustom(method = c("pearson", "kendall", "spearman"), ...)
}
\description{
Custom Correlation Matrix - This really isn't useful yet; taken from Poster_plot.R
}
\examples{
CorMatrixCustom(hy.Fish[hy.Fish$Lake=="Arrow", c(3:ncol(hy.Fish))],
yvar=as.numeric(hy.Fish[hy.Fish$Lake=="Arrow",]$Year),
method="spearman")
}

