% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll2UTM.R
\name{ll2utm}
\alias{ll2utm}
\title{Convert lat/longs into UTMS}
\usage{
ll2utm(file_type = "excel", create_csv = TRUE)
}
\arguments{
\item{file_type}{Currently only supporting excel and csv. Defaults to excel}

\item{create_csv}{Export a csv? Defauls to TRUE}
}
\description{
Using a standard input convert lat/longs into UTMS using the readxl and rgdal package
}
\examples{
#x <- ll2utm(create_csv=FALSE)
#ll2utm(create_csv=TRUE)
}
