% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mav.R
\name{mav}
\alias{mav}
\title{Moving average over a continuous variable}
\source{
http://druedin.com/2012/08/11/moving-averages-in-r/
}
\usage{
mav(x, n = 5)
}
\arguments{
\item{x}{Continuous variable over which to apply the moving average}

\item{n}{number of observation over which to take the moving average}
}
\value{
The moving (or rolling) average of x over the n window
}
\description{
Apply a moving average over a variable defaulting to a window of 5
}

